+++ marlin/src/lib.rs
@@ -67,11 +67,11 @@ macro_rules! eprintln {
 pub mod ahp;
 pub use ahp::*;
 
+/// Implements the Marlin verification gadget.
+pub mod constraints;
+
 /// Implements the base Marlin zkSNARK proof system.
 pub mod marlin;
-
-/// Gadgets for the Marlin SNARK proof system.
-pub mod gadgets;
 
 /// The Marlin public parameters for a given circuit.
 pub mod parameters;
